create useFetchRepositories hook and replace for cleaner code
